Round two of LGTM fix

+++ MekHQ/src/mekhq/campaign/mission/AtBDynamicScenarioFactory.java
@@ -1605,7 +1605,7 @@ public class AtBDynamicScenarioFactory {
             }
         }
 
-        bvBudget += Math.round(bvBudget * scenario.getEffectivePlayerBVMultiplier());
+        bvBudget += (int) Math.round(bvBudget * scenario.getEffectivePlayerBVMultiplier());
 
         // allied bot forces that contribute to BV do not get multiplied by the difficulty
         // even if the player is super good, the AI doesn't get any better
